// src/model/documentoPayload.ts
export interface DocumentoPayload {
    documentoBytes: Blob; 
    ambitoId: number;
    categoriaId: number;
    caserioId: number;
    centroPobladoId: number;
    nombre: string;
    folios: number;
    asunto: string;
    dni: number;
    nombres: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    genero: string;
 }
  